Enhance Windows library linkage and improve DNS entry management.
- Updated CMakeLists.txt to link advapi32 for Windows builds in curl_fo, curl_fo_static, curl_fo_shim, and curl_fo_wrapper targets.
- Introduced reference counting and eviction handling for DNS entries in cache.c, allowing for better memory management and preventing premature deallocation.
- Added proxy handling in curl_dispatch.c and handle.c to support proxy configuration in the shadow structure.
- Improved URL parsing in util.c to correctly handle IPv6 addresses and ports.
- Enhanced error handling in probe.c for socket connection failures.

These changes improve the robustness of the library, particularly in Windows environments, and enhance DNS entry lifecycle management.

+++ src/dns.c
@@ -232,8 +232,8 @@ static int cf_dns_query_type(const char *host, uint16_t qtype,
 
         uint16_t type = (uint16_t)((resp[off] << 8) | resp[off + 1]);
         off += 4;
-        uint32_t ttl = (uint32_t)((resp[off] << 24) | (resp[off+1] << 16) |
-                                  (resp[off+2] << 8) | resp[off+3]);
+        uint32_t ttl = ((uint32_t)resp[off] << 24) | ((uint32_t)resp[off + 1] << 16) |
+                       ((uint32_t)resp[off + 2] << 8) | (uint32_t)resp[off + 3];
         off += 4;
         uint16_t rdlen = (uint16_t)((resp[off] << 8) | resp[off + 1]);
         off += 2;
